requirements.txt
anyio==4.0.0
astor==0.8.1
attrdict==2.0.1
Babel==2.12.1
bce-python-sdk==0.8.90
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
cssselect==1.2.0
cssutils==2.7.1
cycler==0.11.0
Cython==3.0.2
decorator==5.1.1
et-xmlfile==1.1.0
filelock==3.12.4
fire==0.5.0
Flask==2.3.3
flask-babel==3.1.0
flatbuffers==23.5.26
fonttools==4.42.1
future==0.18.3
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
humanfriendly==10.0
idna==3.4
imageio==2.31.4
imgaug==0.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
lmdb==1.4.1
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
networkx==3.1
numpy==1.26.0
onnxruntime==1.16.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
paddle-bfloat==0.1.7
paddleocr==2.7.0.3
paddlepaddle==2.5.1
pandas==2.1.1
pdf2docx==0.5.6
Pillow==10.0.1
premailer==3.10.0
protobuf==3.20.2
psutil==5.9.5
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
pycryptodome==3.19.0
PyMuPDF @ file:///C:/Users/rajee/Python_projects/License_plate_detection_with_OCR/PyMuPDF-1.20.2-cp311-cp311-win_amd64.whl
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
rapidfuzz==3.3.1
rarfile==4.1
requests==2.31.0
scikit-image==0.21.0
scipy==1.11.3
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sympy==1.12
termcolor==2.3.0
tifffile==2023.9.26
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.188
urllib3==2.0.5
visualdl==2.5.3
Werkzeug==2.3.7
